# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Fiber Banking API
  version: 1.0.0
  description: |
    Dokumentasi OpenAPI untuk API perbankan sederhana berbasis Go Fiber.

    Semua pesan response menggunakan bahasa Indonesia dan format standar:
    - Sukses: `{ data: T, status: true, message: string }`
    - Error: `{ status: false, message: string }`
servers:
  - url: http://localhost:3000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: Logout user login (admin, staff, customer)
      description: |
        Mengakhiri sesi di sisi client. Token bearer tetap wajib dikirim agar server dapat memvalidasi sesi aktif.

        role = `admin`, `staff`, `customer`
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
      responses:
        "200":
          description: Logout berhasil
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptySuccessResponse"
              examples:
                success:
                  value:
                    data: null
                    status: true
                    message: Logout berhasil
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/me:
    get:
      tags: [Settings]
      summary: Info user login (admin, staff, customer)
      description: |
        Mengambil informasi user yang sedang login. Jika role customer, response menyertakan info rekening.

        role = `admin`, `staff`, `customer`
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
      responses:
        "200":
          description: Data pengguna berhasil diambil
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeSuccessResponse"
              examples:
                success:
                  value:
                    data:
                      id: ccustomer00000000000000001
                      name: Customer Satu
                      email: customer@example.com
                      role: customer
                      account:
                        account_number: "1000000001"
                        balance: 98750000
                    status: true
                    message: Data pengguna berhasil diambil
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/check-auth:
    get:
      tags: [Settings]
      summary: Cek session valid (admin, staff, customer)
      description: |
        Mengecek apakah bearer token/session saat ini valid.

        role = `admin`, `staff`, `customer`
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
      responses:
        "200":
          description: Session valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckAuthSuccessResponse"
              examples:
                success:
                  value:
                    data:
                      id: cadmin000000000000000001
                      name: Admin Utama
                      email: admin@example.com
                      role: admin
                    message: Session valid
                    status: true
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /api/customer/mutations:
    get:
      tags: [Customer]
      summary: List mutasi customer login (customer)
      description: |
        Mengambil daftar mutasi milik customer yang sedang login.

        role = `customer`
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
        - $ref: "#/components/parameters/SearchQuery"
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/PerPageQuery"
        - $ref: "#/components/parameters/TransactionTypeQuery"
        - $ref: "#/components/parameters/TransactionDirectionQuery"
        - $ref: "#/components/parameters/ActorRoleQuery"
        - $ref: "#/components/parameters/DateFromQuery"
        - $ref: "#/components/parameters/DateToQuery"
      responses:
        "200":
          description: Mutasi berhasil diambil
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MutationListSuccessResponse"
              examples:
                success:
                  value:
                    data:
                      items:
                        - id: ctx000000000000000000001
                          type: transfer
                          direction: out
                          amount: 1250000
                          balance_before: 100000000
                          balance_after: 98750000
                          reference_id: cref000000000000000000001
                          note: Transfer awal antar customer
                          actor_name: Customer Satu
                          actor_role: customer
                          created_at: "2026-07-23 15:30:00"
                      pagination:
                        current_page: 1
                        per_page: 10
                        from: 1
                        total: 1
                        last_page: 1
                    status: true
                    message: Mutasi berhasil diambil
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/customer/mutations/{id}:
    get:
      tags: [Customer]
      summary: Detail mutasi customer login (customer)
      description: Mengambil detail satu mutasi milik customer yang sedang login.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
        - $ref: "#/components/parameters/MutationIdPath"
      responses:
        "200":
          $ref: "#/components/responses/MutationSuccessResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/customer/deposit:
    post:
      tags: [Customer]
      summary: Setor tunai customer login (customer)
      description: |
        Customer melakukan setor tunai ke rekening sendiri.

        role = `customer`
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AmountRequest"
            examples:
              default:
                value:
                  amount: "100000.00"
                  note: Setor tunai mandiri
      responses:
        "201":
          description: Setor tunai berhasil
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationSuccessResponse"
              examples:
                success:
                  value:
                    data:
                      amount: 10000000
                      balance: 110000000
                    status: true
                    message: Setor tunai berhasil
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/customer/withdraw:
    post:
      tags: [Customer]
      summary: Tarik tunai customer login (customer)
      description: |
        Customer melakukan tarik tunai dari rekening sendiri.

        role = `customer`
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AmountRequest"
            examples:
              default:
                value:
                  amount: "50000.00"
                  note: Tarik tunai mandiri
      responses:
        "201":
          description: Tarik tunai berhasil
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationSuccessResponse"
              examples:
                success:
                  value:
                    data:
                      amount: 5000000
                      balance: 105000000
                    status: true
                    message: Tarik tunai berhasil
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/customer/transfer:
    post:
      tags: [Customer]
      summary: Transfer customer login (customer)
      description: |
        Customer melakukan transfer dari rekening sendiri ke customer lain.

        role = `customer`
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerTransferRequest"
            examples:
              default:
                value:
                  to_customer_id: ccustomer00000000000000002
                  amount: "12500.00"
                  note: Transfer belanja
      responses:
        "201":
          description: Transfer berhasil
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationSuccessResponse"
              examples:
                success:
                  value:
                    data:
                      reference_id: cref000000000000000000001
                      amount: 1250000
                      balance: 98750000
                    status: true
                    message: Transfer berhasil
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /api/staff/customers/{customer_id}/deposit:
    post:
      tags: [Staff]
      summary: Setor tunai untuk customer (admin, staff)
      description: |
        Staff atau admin melakukan setor tunai dengan memilih customer.

        role = `admin`, `staff`
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
        - $ref: "#/components/parameters/CustomerIdPath"
      requestBody:
        $ref: "#/components/requestBodies/AmountBody"
      responses:
        "201":
          $ref: "#/components/responses/StaffDepositSuccessResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/staff/customers/{customer_id}/withdraw:
    post:
      tags: [Staff]
      summary: Tarik tunai untuk customer (admin, staff)
      description: |
        Staff atau admin melakukan tarik tunai dengan memilih customer.

        role = `admin`, `staff`
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
        - $ref: "#/components/parameters/CustomerIdPath"
      requestBody:
        $ref: "#/components/requestBodies/AmountBody"
      responses:
        "201":
          $ref: "#/components/responses/StaffWithdrawSuccessResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/staff/customers/{customer_id}/transfer:
    post:
      tags: [Staff]
      summary: Transfer untuk customer (admin, staff)
      description: |
        Staff atau admin melakukan transfer dari customer pilihan ke customer tujuan.

        role = `admin`, `staff`
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
        - $ref: "#/components/parameters/CustomerIdPath"
      requestBody:
        $ref: "#/components/requestBodies/CustomerTransferBody"
      responses:
        "201":
          $ref: "#/components/responses/StaffTransferSuccessResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /api/admin/customers/{customer_id}/deposit:
    post:
      tags: [Admin]
      summary: Admin setor tunai untuk customer (admin)
      description: |
        Alias fitur teller pada prefix admin. Admin melakukan setor tunai dengan memilih customer.

        role = `admin`
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
        - $ref: "#/components/parameters/CustomerIdPath"
      requestBody:
        $ref: "#/components/requestBodies/AmountBody"
      responses:
        "201":
          $ref: "#/components/responses/StaffDepositSuccessResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/admin/customers/{customer_id}/withdraw:
    post:
      tags: [Admin]
      summary: Admin tarik tunai untuk customer (admin)
      description: |
        Alias fitur teller pada prefix admin. Admin melakukan tarik tunai dengan memilih customer.

        role = `admin`
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
        - $ref: "#/components/parameters/CustomerIdPath"
      requestBody:
        $ref: "#/components/requestBodies/AmountBody"
      responses:
        "201":
          $ref: "#/components/responses/StaffWithdrawSuccessResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/admin/customers/{customer_id}/transfer:
    post:
      tags: [Admin]
      summary: Admin transfer untuk customer (admin)
      description: |
        Alias fitur teller pada prefix admin. Admin melakukan transfer dari customer pilihan ke customer tujuan.

        role = `admin`
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AuthorizationHeader"
        - $ref: "#/components/parameters/CustomerIdPath"
      requestBody:
        $ref: "#/components/requestBodies/CustomerTransferBody"
      responses:
        "201":
          $ref: "#/components/responses/StaffTransferSuccessResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Gunakan header `Authorization: Bearer <access_token>`."
  parameters:
    AuthorizationHeader:
      name: Authorization
      in: header
      required: true
      description: Token bearer hasil login.
      schema:
        type: string
        example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.example.signature
    SearchQuery:
      name: q
      in: query
      required: false
      description: Kata kunci pencarian list.
      schema:
        type: string
        example: customer
    PageQuery:
      name: page
      in: query
      required: false
      description: Nomor halaman. Default `1`.
      schema:
        type: integer
        minimum: 1
        default: 1
        example: 1
    PerPageQuery:
      name: per_page
      in: query
      required: false
      description: Jumlah data per halaman. Default `10`, maksimal `100`.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
        example: 10
    CustomerIdPath:
      name: customer_id
      in: path
      required: true
      description: ID user customer tujuan operasi.
      schema:
        type: string
        example: ccustomer00000000000000001
...
    MutationIdPath:
      name: id
      in: path
      required: true
      description: ID mutasi/transaksi.
      schema:
        type: string
        example: ctx000000000000000000001
...
    TransactionTypeQuery:
      name: type
      in: query
      required: false
      description: Filter jenis transaksi.
      schema:
        $ref: "#/components/schemas/TransactionType"
    TransactionDirectionQuery:
      name: direction
      in: query
      required: false
      description: Filter arah transaksi.
      schema:
        $ref: "#/components/schemas/TransactionDirection"
...
    ActorRoleQuery:
      name: actor_role
      in: query
      required: false
      description: Filter role aktor transaksi.
      schema:
        $ref: "#/components/schemas/Role"
    DateFromQuery:
      name: date_from
      in: query
      required: false
      description: Tanggal awal filter mutasi (inklusif).
      schema:
        type: string
        format: date
        example: "2026-07-01"
    DateToQuery:
      name: date_to
      in: query
      required: false
      description: Tanggal akhir filter mutasi (inklusif).
      schema:
        type: string
        format: date
        example: "2026-07-31"
  requestBodies:
    AmountBody:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AmountRequest"
          examples:
            default:
              value:
                amount: "100000.00"
                note: Transaksi teller
    CustomerTransferBody:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CustomerTransferRequest"
          examples:
            default:
              value:
                to_customer_id: ccustomer00000000000000002
                amount: "12500.00"
                note: Transfer via teller
...
  responses:
    ErrorResponse:
      description: Contoh response error standar. Status code aktual dapat berupa 400, 401, 403, 404, 409, 502, atau 500 sesuai kasus.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            error:
              value:
                status: false
                message: Request tidak valid
...
    StaffDepositSuccessResponse:
      description: Setor tunai customer berhasil
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/OperationSuccessResponse"
          examples:
            success:
              value:
                data:
                  amount: 10000000
                  balance: 110000000
                status: true
                message: Setor tunai customer berhasil
    StaffWithdrawSuccessResponse:
      description: Tarik tunai customer berhasil
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/OperationSuccessResponse"
          examples:
            success:
              value:
                data:
                  amount: 5000000
                  balance: 105000000
                status: true
                message: Tarik tunai customer berhasil
    StaffTransferSuccessResponse:
      description: Transfer customer berhasil
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/OperationSuccessResponse"
          examples:
            success:
              value:
                data:
                  transaction_id: ctx000000000000000000001
                  reference_id: cref000000000000000000001
                  amount: 1250000
                  balance: 98750000
                status: true
                message: Transfer customer berhasil
...
    MutationSuccessResponse:
      description: Detail mutasi berhasil diambil
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MutationSuccessResponse"
...
  schemas:
    Role:
      type: string
      enum: [admin, staff, customer]
      example: customer
    TransactionType:
      type: string
      enum: [deposit, withdraw, transfer]
      example: transfer
    TransactionDirection:
      type: string
      enum: [in, out]
      example: out
...
    ErrorResponse:
      type: object
      required: [status, message]
      properties:
        status:
          type: boolean
          example: false
        message:
          type: string
          example: Request tidak valid
    Pagination:
      type: object
      required: [current_page, per_page, from, total, last_page]
      properties:
        current_page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 10
        from:
          type: integer
          example: 1
        total:
          type: integer
          format: int64
          example: 25
        last_page:
          type: integer
          example: 3
...
    AmountRequest:
      type: object
      required: [amount]
      properties:
        amount:
          type: string
          description: Nominal dalam format desimal. Server menyimpan dan mengembalikan nominal sebagai integer cents.
          example: "100000.00"
        note:
          type: string
          nullable: true
          example: Setor tunai mandiri
    CustomerTransferRequest:
      type: object
      required: [to_customer_id, amount]
      properties:
        to_customer_id:
          type: string
          example: ccustomer00000000000000002
        amount:
          type: string
          description: Nominal dalam format desimal. Server menyimpan dan mengembalikan nominal sebagai integer cents.
          example: "12500.00"
        note:
          type: string
          nullable: true
          example: Transfer belanja
...
    UserData:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: string
          example: ccustomer00000000000000001
        name:
          type: string
          example: Customer Satu
        email:
          type: string
          format: email
          example: customer@example.com
        role:
          $ref: "#/components/schemas/Role"
...
    OperationData:
      type: object
      required: [amount, balance]
      properties:
        transaction_id:
          type: string
          nullable: true
          description: ID transaksi utama yang dibuat oleh operasi.
          example: ctx000000000000000000001
        reference_id:
          type: string
          nullable: true
          example: cref000000000000000000001
        amount:
          type: integer
          format: int64
          description: Nominal dalam cents.
          example: 1250000
        balance:
          type: integer
          format: int64
          description: Saldo akhir dalam cents.
          example: 98750000
...
    MutationItem:
      type: object
      required:
        [
          id,
          type,
          direction,
          amount,
          balance_before,
          balance_after,
          actor_name,
          actor_role,
          created_at,
        ]
      properties:
        id:
          type: string
          example: ctx000000000000000000001
        customer_id:
          type: string
          nullable: true
          example: ccustomer00000000000000001
        type:
          $ref: "#/components/schemas/TransactionType"
        direction:
          $ref: "#/components/schemas/TransactionDirection"
        amount:
          type: integer
          format: int64
          example: 1250000
        balance_before:
          type: integer
          format: int64
          example: 100000000
        balance_after:
          type: integer
          format: int64
          example: 98750000
        reference_id:
          type: string
          nullable: true
          example: cref000000000000000000001
        note:
          type: string
          nullable: true
          example: Transfer awal antar customer
        actor_name:
          type: string
          example: Customer Satu
        actor_role:
          $ref: "#/components/schemas/Role"
        created_at:
          type: string
          example: "2026-07-23 15:30:00"
...
    EmptySuccessResponse:
      type: object
      required: [status, message]
      properties:
        data:
          nullable: true
        status:
          type: boolean
          example: true
        message:
          type: string
          example: Berhasil
...
    MeSuccessResponse:
      type: object
      required: [data, status, message]
      properties:
        data:
          allOf:
            - $ref: "#/components/schemas/UserData"
            - type: object
              properties:
                account:
                  type: object
                  nullable: true
                  properties:
                    account_number:
                      type: string
                      example: "1000000001"
                    balance:
                      type: integer
                      format: int64
                      example: 98750000
        status:
          type: boolean
          example: true
        message:
          type: string
          example: Data pengguna berhasil diambil
    CheckAuthSuccessResponse:
      type: object
      required: [data, status, message]
      properties:
        data:
          $ref: "#/components/schemas/UserData"
        status:
          type: boolean
          example: true
        message:
          type: string
          example: Session valid
...
    OperationSuccessResponse:
      type: object
      required: [data, status, message]
      properties:
        data:
          $ref: "#/components/schemas/OperationData"
        status:
          type: boolean
          example: true
        message:
          type: string
          example: Transfer berhasil
    MutationListSuccessResponse:
      type: object
      required: [data, status, message]
      properties:
        data:
          type: object
          required: [items, pagination]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/MutationItem"
            pagination:
              $ref: "#/components/schemas/Pagination"
        status:
          type: boolean
          example: true
        message:
          type: string
          example: Mutasi berhasil diambil
...
    MutationSuccessResponse:
      type: object
      required: [data, status, message]
      properties:
        data:
          $ref: "#/components/schemas/MutationItem"
        status:
          type: boolean
          example: true
        message:
          type: string
          example: Detail mutasi berhasil diambil
...